refactor: Code and Tests
- Refactor code and tests
   - Create TransitionSettingsModal class

Issue #9
